refactor(gui): 治理批 — broad-except-continue 静音清零 + 新棘轮 + 杂项清理
- check_ai_boundaries 新增 --max-except-continue 棘轮：AST 检测 body
  仅为 continue 的宽泛 except（bare/Exception/BaseException）；窄捕获
  （except KeyError: continue）属惯用写法不计入
- 存量 19 处 S112 全部补 logger.debug(..., exc_info=True) 清零
- pyproject 补 ocr/voice extras 依赖声明（Pillow/PyMuPDF、
  SpeechRecognition/PyAudio）与 ai_refactor_contract 文档同步
- ~230 文件杂项清理（简化表达式、import、测试同步）

+++ tool/gui/src/backend/grounded_stats.py
@@ -144,7 +144,7 @@ def format_coverage_line(stats: dict[str, Any]) -> str:
     in_pool = int(stats.get("in_pool", 0))
     outside = int(stats.get("outside_pool", 0))
     new_tagged = int(stats.get("new_tagged", 0))
-    pct = int(round(float(stats.get("coverage_ratio", 0.0)) * 100))
+    pct = round(float(stats.get("coverage_ratio", 0.0)) * 100)
     line = f"池内命中 {in_pool}/{total}（{pct}%）"
     if outside:
         line += f" · 池外 {outside}"
